Renamed file to binaryProperty.ts.
