+++ src/3dTiles/binaryProperty.ts
